requirements.txt
Jinja2==3.1.3
pillow==10.2.0
requests==2.31.0